+++ packages/main/src/Menu.ts
@@ -11,7 +11,6 @@ import {
 } from "@ui5/webcomponents-base/dist/Keys.js";
 import {
 	isPhone,
-	isTablet,
 	isDesktop,
 } from "@ui5/webcomponents-base/dist/Device.js";
 import { getI18nBundle } from "@ui5/webcomponents-base/dist/i18nBundle.js";
@@ -286,13 +285,6 @@ class Menu extends UI5Element {
 	_currentItems!: Array<CurrentItem>;
 
 	/**
-	 * Stores a list of parent menu items for each sub-menu (on phone).
-	 * @private
-	 */
-	@property({ type: Object, multiple: true })
-	_parentItemsStack!: Array<MenuItem>;
-
-	/**
 	 * Stores the ResponsivePopover instance
 	 */
 	@property({ type: Object, defaultValue: undefined })
@@ -366,7 +358,7 @@ class Menu extends UI5Element {
 	}
 
 	get isSubMenuOpened() {
-		return !!this._parentMenuItem;
+		return this._parentMenuItem && this._popover?.isOpen();
 	}
 
 	get menuHeaderTextPhone() {
@@ -374,7 +366,7 @@ class Menu extends UI5Element {
 	}
 
 	onBeforeRendering() {
-		!isPhone() && this._prepareCurrentItems(this.items);
+		this._prepareCurrentItems(this.items);
 
 		const itemsWithChildren = this.itemsWithChildren;
 		const itemsWithIcon = this.itemsWithIcon;
@@ -404,7 +396,7 @@ class Menu extends UI5Element {
 
 		if (this.open) {
 			const opener = this.getOpener();
-			if (opener) {
+			if (opener && !this.isSubMenuOpened) {
 				this.showAt(opener);
 			}
 		} else {
@@ -418,17 +410,14 @@ class Menu extends UI5Element {
 	 * @public
 	 */
 	async showAt(opener: HTMLElement): Promise<void> {
-		if (isPhone()) {
-			this._prepareCurrentItems(this.items);
-			this._parentItemsStack = [];
-		}
 		if (!this._isSubMenu) {
 			this._parentMenuItem = undefined;
 			this._opener = undefined;
 		}
+		const busyWithoutItems = !this._parentMenuItem?.items.length && this._parentMenuItem?.busy;
 		const popover = await this._createPopover();
 		popover.initialFocus = `${this._id}-menu-item-0`;
-		popover.showAt(opener);
+		popover.showAt(opener, busyWithoutItems);
 	}
 
 	/**
@@ -436,17 +425,14 @@ class Menu extends UI5Element {
 	 * @public
 	 */
 	close(): void {
-		if (this._popover) {
-			if (isPhone()) {
-				this._parentItemsStack = [];
-			}
-			this._popover.close(false, false, true);
-		}
+		this._popover?.close(false, false, true);
 	}
 
 	async _createPopover() {
-		const staticAreaItemDomRef = await this.getStaticAreaItemDomRef();
-		this._popover = staticAreaItemDomRef!.querySelector<ResponsivePopover>("[ui5-responsive-popover]")!;
+		if (!this._popover) {
+			const staticAreaItemDomRef = await this.getStaticAreaItemDomRef();
+			this._popover = staticAreaItemDomRef!.querySelector<ResponsivePopover>("[ui5-responsive-popover]")!;
+		}
 		return this._popover;
 	}
 
@@ -456,14 +442,12 @@ class Menu extends UI5Element {
 	}
 
 	_navigateBack() {
-		const parentMenuItem = this._parentItemsStack.pop();
+		this._closeItemSubMenu(this._parentMenuItem as MenuItem, true);
+	}
 
-		this.focus();
-		if (parentMenuItem) {
-			const parentMenuItemParent = parentMenuItem.parentElement as MenuItem;
-			this._prepareCurrentItems(parentMenuItemParent.items);
-			this._parentMenuItem = this._parentItemsStack.length ? this._parentItemsStack[this._parentItemsStack.length - 1] : undefined;
-		}
+	_closeAll() {
+		const mainMenu = this._findMainMenu(this);
+		mainMenu?.close();
 	}
 
 	_prepareCurrentItems(items: Array<MenuItem>) {
@@ -477,6 +461,9 @@ class Menu extends UI5Element {
 	}
 
 	_createSubMenu(item: MenuItem, opener: HTMLElement) {
+		if (item._subMenu) {
+			return;
+		}
 		const ctor = this.constructor as typeof Menu;
 		const subMenu = document.createElement(ctor.getMetadata().getTag()) as Menu;
 
@@ -505,7 +492,7 @@ class Menu extends UI5Element {
 
 	_openItemSubMenu(item: MenuItem, opener: HTMLElement) {
 		const mainMenu = this._findMainMenu(item);
-		mainMenu.fireEvent<MenuBeforeOpenEventDetail>("before-open", {
+		mainMenu?.fireEvent<MenuBeforeOpenEventDetail>("before-open", {
 			item,
 		}, false, false);
 		item._subMenu!.showAt(opener);
@@ -531,8 +518,6 @@ class Menu extends UI5Element {
 
 			if (forceClose || !parentItem._preventSubMenuClose) {
 				subMenu.close();
-				subMenu.remove();
-				parentItem._subMenu = undefined;
 				if (keyboard) {
 					subMenu._opener?.focus();
 				}
@@ -542,7 +527,7 @@ class Menu extends UI5Element {
 		}
 	}
 
-	_prepareSubMenuDesktopTablet(item: MenuItem, opener: HTMLElement) {
+	_prepareSubMenu(item: MenuItem, opener: HTMLElement) {
 		if (opener.id !== this._subMenuOpenerId || (item && item.hasSubmenu)) {
 			// close opened sub-menu if there is any opened
 			this._closeItemSubMenu(this._openedSubMenuItem!, true);
@@ -555,12 +540,6 @@ class Menu extends UI5Element {
 		if (this._parentMenuItem) {
 			this._parentMenuItem._preventSubMenuClose = true;
 		}
-	}
-
-	_prepareSubMenuPhone(item: MenuItem) {
-		this._prepareCurrentItems(item.items);
-		this._parentMenuItem = item;
-		this._parentItemsStack.push(item);
 	}
 
 	_onfocusin(e: FocusEvent): void {
@@ -578,7 +557,7 @@ class Menu extends UI5Element {
 
 		// Sets the new timeout
 		this._timeout = setTimeout(() => {
-			this._prepareSubMenuDesktopTablet(item, opener);
+			this._prepareSubMenu(item, opener);
 		}, MENU_OPEN_DELAY);
 	}
 
@@ -637,7 +616,7 @@ class Menu extends UI5Element {
 			const opener = e.target as OpenerStandardListItem;
 			const item = opener.associatedItem;
 
-			item.hasSubmenu && this._prepareSubMenuDesktopTablet(item, opener);
+			item.hasSubmenu && this._prepareSubMenu(item, opener);
 		} else if (shouldCloseMenu && this._isSubMenu && this._parentMenuItem) {
 			const parentMenuItemParent = this._parentMenuItem.parentElement as Menu;
 			parentMenuItemParent._closeItemSubMenu(this._parentMenuItem, true, true);
@@ -651,9 +630,6 @@ class Menu extends UI5Element {
 		if (!item.hasSubmenu) {
 			// click on an item that doesn't have sub-items fires an "item-click" event
 			if (!this._isSubMenu) {
-				if (isPhone()) {
-					this._parentMenuItem = undefined;
-				}
 				// fire event if the click is on top-level menu item
 				const prevented = !this.fireEvent<MenuItemClickEventDetail>("item-click", {
 					"item": item,
@@ -683,22 +659,22 @@ class Menu extends UI5Element {
 					mainMenu._popover!.close();
 				}
 			}
-		} else if (isPhone()) {
-			// prepares and opens sub-menu on phone
-			this._prepareSubMenuPhone(item);
-		} else if (isTablet()) {
-			// prepares and opens sub-menu on tablet
-			this._prepareSubMenuDesktopTablet(item, opener);
+		} else {
+			this._prepareSubMenu(item, opener);
 		}
 	}
 
-	_findMainMenu(item: MenuItem) {
-		let parentMenu = item.parentElement as Menu;
-		while (parentMenu._parentMenuItem) {
-			parentMenu = parentMenu._parentMenuItem.parentElement as Menu;
+	_findMainMenu(element: MenuItem | Menu) {
+		let menu = this._isMenu(element) ? element as Menu : element.parentElement as Menu;
+		while (menu && menu._parentMenuItem) {
+			menu = menu._parentMenuItem.parentElement as Menu;
 		}
 
-		return parentMenu;
+		return menu;
+	}
+
+	_isMenu(element: HTMLElement) {
+		return element.hasAttribute("ui5-menu");
 	}
 
 	_beforePopoverOpen(e: CustomEvent) {