+++ packages/website/build-scripts/local-cdn.mjs
@@ -19,6 +19,8 @@ await Promise.all([
     cp("../theming/package.json", "local-cdn/local-cdn/theming/package.json"),
     cp("../localization/package.json", "local-cdn/local-cdn/localization/package.json"),
 
+    cp("../../node_modules/@zxing/library/umd/index.min.js", "local-cdn/local-cdn/zxing/umd/index.min.js"),
+    cp("../../node_modules/@zxing/library/esm5/index.js", "local-cdn/local-cdn/zxing/esm5/index.js"),
 ]);
 
 const files = await readdir("local-cdn", {recursive: true, withFileTypes: true});