+++ packages/website/src/components/Editor/index.js
@@ -6,6 +6,15 @@ import useBaseUrl from '@docusaurus/useBaseUrl';
 import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
 import styles from "./index.module.css";
 import { ThemeContext, ContentDensityContext, TextDirectionContext } from "@site/src/theme/Root";
+import { encodeToBase64, decodeFromBase64 } from "./share.js";
+import clsx from "clsx";
+import ShareIcon from "../../../local-cdn/local-cdn/icons/dist/v5/share-2.svg";
+import { Splitter } from 'react-splitter-light';
+import DownloadIcon from "../../../local-cdn/local-cdn/icons/dist/v5/download-from-cloud.svg";
+import EditIcon from "../../../local-cdn/local-cdn/icons/dist/v5/edit.svg";
+import ActionIcon from "../../../local-cdn/local-cdn/icons/dist/v5/action.svg";
+import HideIcon from "../../../local-cdn/local-cdn/icons/dist/v5/hide.svg";
+import downloadSample from './download.js';
 
 if (ExecutionEnvironment.canUseDOM) {
   require('playground-elements');
@@ -27,7 +36,7 @@ const returnProjectToPool = (project) => {
     projectPool.push(project);
 }
 
-export default function Editor({html, js, css, mainFile = "main.js" }) {
+export default function Editor({html, js, css, mainFile = "main.js", canShare = false, standalone = false, mainFileSelected = false }) {
   const projectContainerRef = useRef(null);
   const projectRef = useRef(null);
   const previewRef = useRef(null);
@@ -38,11 +47,11 @@ export default function Editor({html, js, css, mainFile = "main.js" }) {
   // name is set on iframe so it can be passed back in resize message to identify which iframe is resized
   const iframeId = useId();
   const [editorVisible, setEditorVisible] = useState(false);
-  const [btnText, setButtonText] = useState("Edit");
   const {siteConfig, siteMetadata} = useDocusaurusContext();
   const { theme, setTheme } = useContext(ThemeContext);
   const { contentDensity, setContentDensity } = useContext(ContentDensityContext);
   const { textDirection, setTextDirection } = useContext(TextDirectionContext);
+  const [copied, setCopied] = useState(false);
 
   function addImportMap(html) {
     return html.replace("<head>", `
@@ -56,10 +65,17 @@ export default function Editor({html, js, css, mainFile = "main.js" }) {
           "@ui5/webcomponents-localization/": "${getHostBaseUrl()}local-cdn/localization/",
           "@ui5/webcomponents-theming/": "${getHostBaseUrl()}local-cdn/theming/",
           "lit-html": "${getHostBaseUrl()}local-cdn/lit-html/lit-html.js",
-          "lit-html/": "${getHostBaseUrl()}local-cdn/lit-html/"
+          "lit-html/": "${getHostBaseUrl()}local-cdn/lit-html/",
+          "@zxing/library/umd/": "${getHostBaseUrl()}local-cdn/zxing/umd/",
+          "@zxing/library/esm5/": "${getHostBaseUrl()}local-cdn/zxing/esm5/"
         }
       }
     </script>
+    <style>
+      *:not(:defined) {
+        display: none;
+      }
+    </style>
 `)
   }
 
@@ -84,14 +100,65 @@ export default function Editor({html, js, css, mainFile = "main.js" }) {
 
   function toggleEditor() {
     setEditorVisible(!editorVisible);
-    setButtonText(editorVisible ? "Edit" : "Hide code");
+  }
+
+  const getSampleFiles = () => {
+    const files = {};
+
+    // convert file format
+    projectRef.current.files.forEach(f => {
+      files[f.name] = {
+        name: f.name,
+        content: f.content
+      };
+    });
+
+    // remove import map from index.html
+    const htmlContent = files["index.html"].content;
+    const startIdx = htmlContent.indexOf(`<script type="importmap">`);
+    const endIdx = htmlContent.indexOf(`</script>`) + `</script>`.length;
+    files["index.html"].content = htmlContent.substring(0, startIdx) + htmlContent.substring(endIdx)
+
+    // remove playground support
+    delete files["playground-support.js"];
+
+    return files;
+  }
+
+  const download = () => {
+    const files = getSampleFiles();
+    downloadSample(files);
+  }
+
+  const share = () => {
+    const files = getSampleFiles();
+
+    // encode and put in url
+    const hash = encodeToBase64(JSON.stringify(files));
+    navigator.clipboard.writeText(new URL(`#${hash}`, window.location.href).href);
+    setCopied(true);
+  }
+
+  const saveProject = () => {
+    const files = getSampleFiles();
+    localStorage.setItem("project", JSON.stringify(files));
   }
 
   const baseUrl = useBaseUrl("/");
+  const playUrl = useBaseUrl("/play");
+
+  const openInNewTab = () => {
+    const files = getSampleFiles();
+
+    // encode and put in url
+    const hash = encodeToBase64(JSON.stringify(files));
+    const url = new URL(`${playUrl}#${hash}`, location.origin);
+    window.open(url, "_blank");
+  }
 
   useEffect(() => {
     projectRef.current = getProjectFromPool();
-    projectRef.current.config = {
+    let newConfig = {
       files: {
         "index.html": {
           content: addImportMap(fixAssetPaths(html)),
@@ -104,7 +171,8 @@ export default function Editor({html, js, css, mainFile = "main.js" }) {
           content: `/* playground-hide */
 import "./playground-support.js";
 /* playground-hide-end */
-${fixAssetPaths(js)}`
+${fixAssetPaths(js)}`,
+          selected: mainFileSelected,
         },
         "main.css": {
           content: css,
@@ -122,6 +190,42 @@ ${fixAssetPaths(js)}`
         }
       }
     }
+    if (newConfig.files["main.css"].hidden) {
+      delete newConfig.files["main.css"];
+    }
+
+    // restore project if saved
+    if (location.pathname.endsWith("/play") || location.pathname.endsWith("/play/")) {
+      const savedProject = localStorage.getItem("project");
+      if (savedProject) {
+        try {
+          const savedConfig = JSON.parse(savedProject);
+          savedConfig["index.html"].content = addImportMap(fixAssetPaths(savedConfig["index.html"].content));
+          if (savedConfig["main.js"] && newConfig.files["main.ts"]) {
+            delete newConfig.files["main.ts"];
+          }
+          newConfig.files = {...newConfig.files, ...savedConfig};
+        } catch (e) {
+          console.log(e);
+        }
+      }
+    }
+
+    // shared content - should be after restore from localstorage
+    if ((location.pathname.endsWith("/play") || location.pathname.endsWith("/play/")) && location.hash) {
+      try {
+        const sharedConfig = JSON.parse(decodeFromBase64(location.hash.replace("#", "")));
+        sharedConfig["index.html"].content = addImportMap(fixAssetPaths(sharedConfig["index.html"].content));
+        if (sharedConfig["main.js"] && newConfig.files["main.ts"]) {
+          delete newConfig.files["main.ts"];
+        }
+        newConfig.files = {...newConfig.files, ...sharedConfig};
+      } catch (e) {
+        console.log(e);
+      }
+    }
+
+    projectRef.current.config = newConfig;
     projectContainerRef.current.appendChild(projectRef.current)
 
     const messageHandler = async (event) => {
@@ -131,15 +235,21 @@ ${fixAssetPaths(js)}`
     }
     window.addEventListener("message", messageHandler);
 
-    previewRef.current.project = projectRef.current;
     tabBarRef.current.project = projectRef.current;
     fileEditorRef.current.project = projectRef.current;
+    previewRef.current.project = projectRef.current;
 
     tabBarRef.current.editor = fileEditorRef.current;
+
+    // setup localstorage saving
+    if (standalone) {
+      projectRef.current.addEventListener("compileStart", saveProject);
+    }
 
     return function () {
       // component cleanup
       window.removeEventListener("message", messageHandler);
+      projectRef.current.removeEventListener("compileStart", saveProject);
       returnProjectToPool(projectRef.current);
     }
   }, []);
@@ -156,22 +266,141 @@ ${fixAssetPaths(js)}`
     projectRef.current.config = newConfig;
   }, [theme, contentDensity, textDirection]);
 
+  useEffect(() => {
+    if (copied) {
+      setTimeout(() => {
+        setCopied(false);
+      }, 5000)
+    }
+  }, [copied]);
+
+  function optionalSplitter(editor, preview) {
+    return (
+      <>
+        { standalone
+          ?
+            <div style={{width: "100%"}}>
+              <Splitter>
+                {preview}
+                {editor}
+              </Splitter>
+            </div>
+          :
+            <div>
+              {editor}
+              {preview}
+            </div>
+        }
+      </>
+    )
+  }
+
+  function preview() {
+    return (
+      <>
+        <playground-preview class={clsx(styles.previewResultHidden, {
+            [styles['preview-standalone']]: standalone,
+            [styles['preview-sample']]: !standalone,
+          })}
+          style={{ height: "unset", minHeight: "7rem" }} ref={previewRef}
+        ></playground-preview>
+      </>
+    )
+  }
+
+  function editor() {
+    return (
+      <>
+        <div
+          className={clsx({
+            [styles['editor-standalone']]: standalone,
+            [styles['editor-sample']]: !standalone,
+          })}
+          style={{display: editorVisible | standalone ? "block" : "none"}}>
+          <playground-tab-bar editable-file-system ref={tabBarRef}></playground-tab-bar>
+          <playground-file-editor line-numbers ref={fileEditorRef}></playground-file-editor>
+        </div>
+      </>
+    )
+  }
+
   return (
     <>
       <div ref={projectContainerRef}></div>
-      <div style={{display: "flex", flexDirection: "column", border: "1px solid hsla(203, 50%, 30%, 0.15)", boxShadow: "var(--ifm-color-secondary) 0 0 3px 0", borderRadius: "0.5rem", overflow: "hidden" }}>
-        <playground-preview class={ styles.previewResultHidden } style={{ height: "unset", minHeight: "7rem" }} ref={previewRef}></playground-preview>
-          <div style={{display: editorVisible ? "block" : "none"}}>
-            <playground-tab-bar editable-file-system ref={tabBarRef}></playground-tab-bar>
-            <playground-file-editor line-numbers ref={fileEditorRef}></playground-file-editor>
-          </div>
-          <button
-            className={"button " + (editorVisible ? "button--secondary" : "button--primary")}
-            style={{ borderEndEndRadius: 0, borderTopRightRadius:0, padding: "0.125rem 0.75rem", margin: "0", alignSelf: "end", fontSize: "0.625rem" }}
-            onClick={ toggleEditor }
-          >
-            {btnText}
-          </button>
+
+      {canShare
+        ?
+          <>
+            <div className={`${styles.editor__toolbar}`}>
+              <button
+                className={`button button--secondary ${styles.previewResult__download}`}
+                onClick={ download }
+              >
+               <DownloadIcon className={`${styles.btn__icon}`}/>
+                Download
+              </button>
+
+              <button
+                className={`button button--secondary ${styles.previewResult__share}`}
+                onClick={ share }
+              >
+               <ShareIcon className={`${styles.btn__icon}`}/>
+                Share
+              </button>
+
+              { copied
+                ? <div style={ {position: "absolute"} }>
+                    <span className={styles["copy-status"]}>&#x2714; Link copied</span>
+                  </div>
+                : <></>
+              }
+            </div>
+          </>
+        :
+          <></>
+      }
+
+      <div
+        className={clsx({
+          [styles['container-standalone']]: standalone,
+          [styles['container-sample']]: !standalone,
+        })}
+        style={{ border: "1px solid hsla(203, 50%, 30%, 0.15)", boxShadow: "var(--ifm-color-secondary) 0 0 3px 0", borderRadius: "0.5rem", overflow: "hidden" }}
+      >
+        {optionalSplitter(preview(), editor())}
+        <div className={ `${styles.previewResult__actions}  ${(canShare ? styles.previewResult__hasShare : "")} `}>
+        { standalone
+          ?
+            <></>
+          :
+          <>
+            <button
+              className={`button button--secondary ${styles.previewResult__downloadSample}`}
+              onClick={ download }
+            >
+              <DownloadIcon className={`${styles["btn__icon--edit"]} `}/>
+              Download
+            </button>
+
+            <button
+              className={`button button--secondary ${styles.previewResult__downloadSample}`}
+              onClick={ openInNewTab }
+            >
+              <ActionIcon className={`${styles["btn__icon--edit"]} `}/>
+              Open in Playground
+            </button>
+
+            <button
+              className={`button ${(editorVisible ? "button--secondary" : "button--secondary")} ${styles.previewResult__toggleCodeEditor} ${(canShare ? styles.previewResult__hasShare : "")}` }
+              onClick={ toggleEditor }
+            >
+              {editorVisible ? <HideIcon className={`${styles["btn__icon--edit"]} `}/> : <EditIcon className={`${styles["btn__icon--edit"]}`}/>}
+              {editorVisible ? "Hide code" : "Edit"}
+            </button>
+          </>
+        }
+        </div>
+
       </div>
     </>
   );